fix: camera ip select
