+++ mobile/src/components/selects/select/Select.tsx
@@ -1,6 +1,7 @@
 import { IonLabel, IonSelect, IonSelectOption } from "@ionic/react";
 import "./Select.scss";
 import { SelectItem } from "../../../models/types/selectItem";
+import { useTranslation } from "react-i18next";
 
 type SelectProps = {
   value: string;
@@ -8,25 +9,34 @@ type SelectProps = {
   placeholder: string;
   selectList: SelectItem[];
   handleChange: (e: any) => void;
+  handleFocus?: () => void;
+  isLoading?: boolean;
 };
 
-const Select = ({ label, placeholder, selectList, value, handleChange }: SelectProps) => {
+const Select = ({ label, placeholder, selectList, value, handleChange, handleFocus, isLoading}: SelectProps) => {
+  const {t} = useTranslation();
+
   return (
     <div className="select__wrapper">
       <IonLabel className="select__label label label__bold">{label}</IonLabel>
       <IonSelect
         onIonChange={handleChange}
+        onIonFocus={handleFocus}
         value={value}
         labelPlacement="stacked"
         interface="popover"
         placeholder={placeholder}
         className="select"
       >
-        {selectList.map(({ label, value }) => (
+
+        {isLoading ? 
+          <IonSelectOption disabled>{t("messages.loading")}...</IonSelectOption> 
+          : (selectList.length > 0 ? selectList.map(({ label, value }) => (
           <IonSelectOption key={value} value={value}>
             {label}
-          </IonSelectOption>
-        ))}
+          </IonSelectOption>)) 
+          : <IonSelectOption disabled>{t("messages.noData")}</IonSelectOption>)
+        }
       </IonSelect>
     </div>
   );
