+++ mobile/src/locales/ru/translation.json
@@ -19,6 +19,8 @@
     "erp": "Подключения ERP"
   },
   "messages": {
+    "noData": "Ничего не найдено",
+    "loading": "Загрузка",
     "noDatabases": "Нет доступных баз данных",
     "noReports": "Не удалось найти запрашиваемый отчет",
     "noCameras": "Не удалось найти камеры для этой операции",
