+++ mobile/src/pages/cameras/add-camera/AddCamera.tsx
@@ -1,34 +1,26 @@
 import { useTranslation } from "react-i18next";
 import { ROUTES } from "../../../shared/constants/routes";
 import { useHistory } from "react-router-dom";
-import { useEffect, useState } from "react";
+import { useState } from "react";
 import { useCookies } from "react-cookie";
 import { Preloader } from "../../../components/preloader/preloader";
 import {
   IonButton,
   IonContent,
   IonPage,
-  IonSegment,
-  IonSegmentButton,
-  IonText,
   IonToast,
-  useIonViewWillEnter,
 } from "@ionic/react";
 import { Header } from "../../../components/header/Header";
 import { ConfirmationModal } from "../../../components/confirmationModal/confirmationModal";
-import BottomButton from "../../../components/bottomButton/BottomButton";
 import { TOAST_DELAY } from "../../../constants/toastDelay";
 import CameraSegment from "../../../components/cameraSegment/cameraSegment";
-import { findCamera, getSelectedCameras } from "../../../api/cameraRequest";
 import { postAlgorithnDependences } from "../../../api/algorithmRequest";
-import Zones from "../../../components/zoneSegment/zones/zones";
-import { SelectItem } from "../../../models/types/selectItem";
 
 const AddCamera = () => {
   const { t } = useTranslation();
   const history = useHistory();
   const [cookies] = useCookies(["token"]);
-  const [loading, setLoading] = useState(true);
+  const [loading, setLoading] = useState(false);
   const [isOpenModal, setIsOpenModal] = useState(false);
   const [toastMessage, setToastMessage] = useState("");
   const [selectedSegment, setSelectedSegment] = useState<"camera" | "zone">("camera");
@@ -40,24 +32,8 @@ const AddCamera = () => {
   const [password, setPassword] = useState("");
   const [isNotification, setIsNotification] = useState(false);
   const [cameraSelect, setCameraSelect] = useState(false);
-  const [isCreateCamera, setIsCreateCamera] = useState(false);
-  const [findCameraList, setFindCameraList] = useState([]);
   const [isNotificationAfterCreate, setIsNotificationAfterCreate] = useState(false);
-  const [createdCameras, setCreatedCameras] = useState<any>([]);
-  const [error, setError] = useState(false);
 
-  useIonViewWillEnter(() => {
-    getSelectedCameras(window.location.hostname, cookies.token)
-      .then(response => {
-        let cameras = [];
-        if (response.data.length > 0) {
-          cameras = response.data.sort((a, b) => a.name.localeCompare(b.name));
-          setCreatedCameras(cameras);
-        }
-        showAddCameras(cameras);
-      })
-      .catch(error => setError(error.message));
-  });
 
   const isBlank = !(cameraIP && userName && password);
 
@@ -86,26 +62,6 @@ const AddCamera = () => {
   const handleSave = () => {
     setIsOpenModal(false);
     applySettings();
-  };
-
-  const showAddCameras = cameras => {
-    findCamera(window.location.hostname)
-      .then(response => {
-        if (response.data && response.data.results) {
-          const allCameras = response.data.results;
-          const bufCreatedCameras = cameras.length > 0 ? cameras.map(e => e.id) : [];
-          const resultCameras = allCameras.filter(value => {
-            return !bufCreatedCameras.includes(value);
-          });
-          setFindCameraList(resultCameras);
-        } else {
-          setFindCameraList([]);
-        }
-      })
-      .catch(error => console.log(error.message))
-      .finally(() => setLoading(false));
-    setIsCreateCamera(true);
-    setCameraSelect(true);
   };
 
   const applySettings = async () => {
@@ -162,14 +118,11 @@ const AddCamera = () => {
               {/* {selectedSegment === "camera" && ( */}
               <CameraSegment
                 cameraIP={cameraIP}
-                isCreateCamera={isCreateCamera}
-                cameraSelect={cameraSelect}
                 setCameraIP={ip => setCameraIP(ip)}
                 userName={userName}
                 password={password}
                 applySettings={applySettings}
                 isEnabled={isEnabled}
-                findCameraList={findCameraList}
                 cameraName={cameraName}
                 setUserName={name => setUserName(name)}
                 setPassword={password => setPassword(password)}
